+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to
 reference date, if available (e.g. live streams)
 - `Live` indicator for YouTube live streams in post-annotator
 - Edit media screen allows adding duration for live streams once archived
+- Duration field on edit media screen
 
 ### Changed